Adding `--dry-run` and `--path` CLI option (#60)
* (mod/main)adding --path flag

This flag should help the user specify where should le scan result be stored (this will not be appended to `.yelaa`; it is more intuitive to let the user chose the whole path

* (mod/tools)adapting interface to accept scanPath option

This commit adds the scanPath option where necessary for every tool that requires an out-directory. Extensive testing must still be done, but everything seems to work so far.

* (fix/gb)gobuster running rule

GoBuster should be excluded from ToolList since we configure it manually.

It should also run outside of the loop iterating over all tools, for obvious reasons.

* (fix/tool)assetfinder configuration

assetfinder functions config would break the initalisation of scan.

* (mod/main)support for --dryRun option

adding support for dry run option from the CLI
also fixed an issue where gobuster dir would not get created if its parent didnt exist

* (fix/main)creating out-dir only after CLI parsing

Since we need the directory  name after it has been parsed, we run `createOutDirectory()` in the `Run()` functions of the commands parsed by Cobra. Otherwise, we would have always gotten the default interface, which is not the desired behaviour.