(fix/proxy)using GetHttpTransport() to configure http settings
still a some work to do in order to have Proxy set for every tool if
it is specified by the user, but having a helper function to easily
fetch http.Transport with the correct values is a good start

(also fixed some english spelling mistakes)

next step is to add the proxy to the configuration of bigger projects,
like httpx & nuclei. AFAIK httpx does not use ProxyFromEnvironment in
order to set the proxy, which means we probably will have to set it
manually. This can be annoying as this is not a mandatory setting

will also probably need to rework dorks through a cleaner interface,
because using exec.Command() prevents us from manually setting a proxy.