(mod/dorks)using dorks package instead of google-chrome browser
yesterday, i wrote a little go package to run dorks and retrieve files,
so that we no longer need to have the chrome browser installed.

this will greatly reduce the size of the docker image, and allow us to
use a proxy to run dorks, which is cool.

there is still a lot of room for improvements for the dork package but
this is a good start, i guess