(fix/tools)proxy usage with socks & http (#69)
* (fix/proxy)using GetHttpTransport() to configure http settings

still a some work to do in order to have Proxy set for every tool if
it is specified by the user, but having a helper function to easily
fetch http.Transport with the correct values is a good start

(also fixed some english spelling mistakes)

next step is to add the proxy to the configuration of bigger projects,
like httpx & nuclei. AFAIK httpx does not use ProxyFromEnvironment in
order to set the proxy, which means we probably will have to set it
manually. This can be annoying as this is not a mandatory setting

will also probably need to rework dorks through a cleaner interface,
because using exec.Command() prevents us from manually setting a proxy.

* (mod/build)adding rule to dynamically build Yelaa

To run Yelaa with Proxychains, one needs to have a dynamically compiled
program. This is because Proxychains uses LD_PRELOAD tricks to hook to
connect() and set the proxy, which is not possible with a statically
program such as one compile with gc (standard go compiler)

the script used to compile is still very much a WIP and other options
could be explored instead:
    - use a TUN interface (https://github.com/nicocha30/ligolo-ng)
    - TUN to SOCKS (https://github.com/russdill/tunsocks)
    - iptables tricks (https://github.com/n1nj4sec/pr0cks) but this
      would not work very well with docker

* (mod/gowitness)setting proxy option in chrome

- dorks was broken with socksproxy so the option for that was removed.

scan command is broken, further testing is required

* (fix/tools)manually setting proxy for most interfaces

* (mod/docs)improving docs regarding make dynamic rule

* (mod/main)removing YELAA_PROXY env value usage

* (mod/ci)bumping tool version & improving docs a little